Initial commit
To be perfectly honest this isn't the first commit but so far all
of the commits have been either nonsensical or incomprehensible
("fixed things", "oops", "now working" and so on) it makes sense
to just squash everything into one big commit.

The next commits will be a lot more interesting. I promise.