Remove token refresh from storage interface
The interface should be a pure storage interface and the token
refresh code should be independent of the actual *store*.

Move that code out and add two new methods for storage interface.

Added a build tag for the connect test since it uses a local HTTP
test server on port 8080 which doesn't work well for build servers.

Test with go test -tags testlocally if you want to run the tests.