Create tables with IF NOT EXISTS
Make DDL a bit nicer to use with existing databases; it is now possible
to set the create schema parameter every time the storage is initialised.

This could - in theory - create issues if the existing structure is
different from the expected but it will be caught during startup
anyways sice we're using prepared statements.